# Update queries to use aggregated data