--- a/.env.example
+++ b/.env.example
@@ -88,6 +88,7 @@ MAIL_FROM_NAME="${APP_NAME}"
 # Feature Flags
 
 FEATURE_BEAT=true
+FEATURE_AGGREGATE_QUERIES=true
 
 # Providers