--- a/app/Helpers/database/game.php
+++ b/app/Helpers/database/game.php
@@ -88,27 +88,29 @@ function getGameMetadata(
         if (env('FEATURE_AGGREGATE_QUERIES')) {
             $parentGameId = getParentGameIdFromGameTitle($gameDataOut['Title'], $gameDataOut['ConsoleID']);
 
-            $gameIdStatement = 'game_id = :gameId';
-            if ($parentGameId !== null) {
+            if ($parentGameId === null) {
+                $query = "SELECT players_total AS NumDistinctPlayers FROM GameData WHERE ID=$gameID";
+                $gameMetrics = legacyDbFetch($query);
+            } else {
                 $bindings['parentGameId'] = $parentGameId;
                 $gameIdStatement = 'game_id IN (:gameId, :parentGameId)';
-            }
 
-            $bindings = ['gameId' => $gameID];
-            $query = "SELECT COUNT(IF(achievements_unlocked > 0, 1, NULL)) AS NumDistinctPlayersCasual,
-                             COUNT(IF(achievements_unlocked_hardcore > 0, 1, NULL)) AS NumDistinctPlayersHardcore
-                        FROM (
-                            SELECT user_id, achievements_unlocked, achievements_unlocked_hardcore
-                              FROM player_games WHERE $gameIdStatement AND achievements_unlocked > 0
-                             GROUP BY user_id
-                        ) AS InnerTable";
-            $gameMetrics = legacyDbFetch($query, $bindings);
+                $bindings = ['gameId' => $gameID];
+                $query = "SELECT COUNT(*) AS NumDistinctPlayers
+                            FROM (
+                                SELECT user_id, SUM(achievements_unlocked) as NumAchievementsUnlocked
+                                FROM player_games WHERE $gameIdStatement
+                                GROUP BY user_id
+                                HAVING NumAchievementsUnlocked > 0
+                            ) AS InnerTable";
+                $gameMetrics = legacyDbFetch($query, $bindings);
+            }
 
             $gameMetrics['ParentGameID'] = $parentGameId;
         } else {
             $cacheKey = "game:$gameID:playercount";
             $gameMetrics = Cache::get($cacheKey);
-            if ($gameMetrics === null) {
+            if ($gameMetrics === null || !array_key_exists('NumDistinctPlayers', $gameMetrics)) {
                 $parentGameId = getParentGameIdFromGameTitle($gameDataOut['Title'], $gameDataOut['ConsoleID']);
 
                 $bindings = [
@@ -122,73 +124,64 @@ function getGameMetadata(
                     $gameIdStatement = 'ach.GameID IN (:gameId, :parentGameId)';
                 }
 
-                $query = "SELECT aw.HardcoreMode, COUNT(DISTINCT aw.User) as Users
+                $query = "SELECT COUNT(DISTINCT aw.User) as Users
                             FROM Awarded AS aw
                             LEFT JOIN Achievements AS ach ON ach.ID = aw.AchievementID
                             LEFT JOIN UserAccounts AS ua ON ua.User = aw.User
                             WHERE $gameIdStatement AND ach.Flags = :achievementFlag
-                            AND NOT ua.Untracked
-                        GROUP BY aw.HardcoreMode";
+                            AND NOT ua.Untracked AND aw.HardcoreMode = 0";
 
-                $gameMetaData = legacyDbFetchAll($query, $bindings);
-
-                $numDistinctPlayersCasual = 0;
-                $numDistinctPlayersHardcore = 0;
-                foreach ($gameMetaData as $data) {
-                    if ($data['HardcoreMode'] == UnlockMode::Hardcore) {
-                        $numDistinctPlayersHardcore = $data['Users'];
-                    } else {
-                        $numDistinctPlayersCasual = $data['Users'];
-                    }
-                }
+                $numDistinctPlayers = legacyDbFetch($query, $bindings)['Users'];
 
                 $gameMetrics = [
                     'ParentGameID' => $parentGameId,
-                    'NumDistinctPlayersCasual' => $numDistinctPlayersCasual,
-                    'NumDistinctPlayersHardcore' => $numDistinctPlayersHardcore,
+                    'NumDistinctPlayers' => $numDistinctPlayers,
                 ];
 
                 // if a game has more than 100 players, only update the player count once an hour
-                if ($numDistinctPlayersCasual > 100) {
+                if ($numDistinctPlayers > 100) {
                     Cache::put($cacheKey, $gameMetrics, Carbon::now()->addHours(1));
                 }
             }
         }
 
         $gameDataOut['ParentGameID'] = $gameMetrics['ParentGameID'];
-        $gameDataOut['NumDistinctPlayersCasual'] = $gameMetrics['NumDistinctPlayersCasual'];
-        $gameDataOut['NumDistinctPlayersHardcore'] = $gameMetrics['NumDistinctPlayersHardcore'];
+        $gameDataOut['NumDistinctPlayers'] = $gameMetrics['NumDistinctPlayers'];
 
-        $metricsBindings = [
-            'metricsGameId' => $gameID,
-            'metricsAchievementFlag' => $flag,
-        ];
-        $metricsColumns = 'IFNULL(tracked_aw.NumAwarded, 0) AS NumAwarded,
-                           IFNULL(tracked_aw.NumAwardedHardcore, 0) AS NumAwardedHardcore,';
-
-        // if a game has more than 100 players, don't filter out the untracked users as the
-        // join becomes very expensive. will be addressed when denormalized data is captured
-        if ($gameMetrics['NumDistinctPlayersCasual'] > 100) {
-            $metricsJoin = "LEFT JOIN (
-                SELECT ach.ID AS AchievementID,
-                    (COUNT(aw.AchievementID) - SUM(IFNULL(aw.HardcoreMode, 0))) AS NumAwarded,
-                    (SUM(IFNULL(aw.HardcoreMode, 0))) AS NumAwardedHardcore
-                FROM Achievements AS ach
-                LEFT JOIN Awarded AS aw ON aw.AchievementID = ach.ID
-                WHERE ach.GameID = :metricsGameId AND ach.Flags = :metricsAchievementFlag
-                GROUP BY ach.ID
-            ) AS tracked_aw ON tracked_aw.AchievementID = ach.ID";
+        if (env('FEATURE_AGGREGATE_QUERIES')) {
+            $metricsColumns = 'ach.unlocks_total AS NumAwarded, ach.unlocks_hardcore_total AS NumAwardedHardcore,';
         } else {
-            $metricsJoin = "LEFT JOIN (
-                SELECT ach.ID AS AchievementID,
-                    (COUNT(aw.AchievementID) - SUM(IFNULL(aw.HardcoreMode, 0))) AS NumAwarded,
-                    (SUM(IFNULL(aw.HardcoreMode, 0))) AS NumAwardedHardcore
-                FROM Achievements AS ach
-                LEFT JOIN Awarded AS aw ON aw.AchievementID = ach.ID
-                LEFT JOIN UserAccounts AS ua ON ua.User = aw.User
-                WHERE ach.GameID = :metricsGameId AND ach.Flags = :metricsAchievementFlag AND NOT ua.Untracked
-                GROUP BY ach.ID
-            ) AS tracked_aw ON tracked_aw.AchievementID = ach.ID";
+            $metricsBindings = [
+                'metricsGameId' => $gameID,
+                'metricsAchievementFlag' => $flag,
+            ];
+            $metricsColumns = 'IFNULL(tracked_aw.NumAwarded, 0) AS NumAwarded,
+                            IFNULL(tracked_aw.NumAwardedHardcore, 0) AS NumAwardedHardcore,';
+
+            // if a game has more than 100 players, don't filter out the untracked users as the
+            // join becomes very expensive. will be addressed when denormalized data is captured
+            if ($gameMetrics['NumDistinctPlayers'] > 100) {
+                $metricsJoin = "LEFT JOIN (
+                    SELECT ach.ID AS AchievementID,
+                        (COUNT(aw.AchievementID) - SUM(IFNULL(aw.HardcoreMode, 0))) AS NumAwarded,
+                        (SUM(IFNULL(aw.HardcoreMode, 0))) AS NumAwardedHardcore
+                    FROM Achievements AS ach
+                    LEFT JOIN Awarded AS aw ON aw.AchievementID = ach.ID
+                    WHERE ach.GameID = :metricsGameId AND ach.Flags = :metricsAchievementFlag
+                    GROUP BY ach.ID
+                ) AS tracked_aw ON tracked_aw.AchievementID = ach.ID";
+            } else {
+                $metricsJoin = "LEFT JOIN (
+                    SELECT ach.ID AS AchievementID,
+                        (COUNT(aw.AchievementID) - SUM(IFNULL(aw.HardcoreMode, 0))) AS NumAwarded,
+                        (SUM(IFNULL(aw.HardcoreMode, 0))) AS NumAwardedHardcore
+                    FROM Achievements AS ach
+                    LEFT JOIN Awarded AS aw ON aw.AchievementID = ach.ID
+                    LEFT JOIN UserAccounts AS ua ON ua.User = aw.User
+                    WHERE ach.GameID = :metricsGameId AND ach.Flags = :metricsAchievementFlag AND NOT ua.Untracked
+                    GROUP BY ach.ID
+                ) AS tracked_aw ON tracked_aw.AchievementID = ach.ID";
+            }
         }
     }