--- a/app/Helpers/database/game.php
+++ b/app/Helpers/database/game.php
@@ -3,7 +3,6 @@
 use App\Community\Enums\ArticleType;
 use App\Community\Enums\TicketState;
 use App\Platform\Enums\AchievementFlag;
-use App\Platform\Enums\UnlockMode;
 use App\Platform\Models\Game;
 use App\Site\Enums\Permissions;
 use Illuminate\Support\Carbon;