New queries are enabled by setting the `FEATURE_AGGREGATE_QUERIES` .env flag to true.

Data populated for testing was done by hand by modifying the code to run the old queries and update the new tables with the resulting data. No new rows were added, but all of the existing rows had uninitialized fields. I did not keep the related code - I apologize to anyone else who wants to test this before luchaos finishes his PR.