@luchaos 2023-09-01
In case you want to include and validate the missing index right away:
```php
// database/migrations/platform/2023_09_01_000000_update_player_achievements_table.php
<?php

declare(strict_types=1);

use Illuminate\Database\Migrations\Migration;
use Illuminate\Database\Schema\Blueprint;
use Illuminate\Support\Facades\Schema;

return new class() extends Migration {
    public function up(): void
    {
        Schema::table('player_achievements', function (Blueprint $table) {
            $table->index('unlocked_hardcore_at');
        });
    }

    public function down(): void
    {
        Schema::table('player_achievements', function (Blueprint $table) {
            $table->dropIndex('unlocked_hardcore_at');
        });
    }
};
```