@Jamiras 2023-09-02
Additional timings for the three disabled APIs:

**API_GetGameInfoAndUserProgress?g=1446&u=Jamiras**
With feature: 210ms
Without feature: 1813ms

**API_GetUserRecentlyPlayedGames?u=Jamiras**
With feature: 373ms
Without feature: 2339ms

**API_GetUserSummary?u=Maxmilyin**
With feature: 1962ms
Without feature: 3423ms